# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/auth/signup": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign up.",
                "description": "User sign up.",
                "operationId": "d2361024a07caa3d3d466ee71a6e8ff3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignupData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/signin": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign in.",
                "description": "User sign in.",
                "operationId": "acb28cacb24a8b6866e68e3be09b90db",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SigninData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JwtTokenResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Refresh.",
                "description": "Refresh a token.",
                "operationId": "a02d175394e9af4b84c7ae1d38ff7f33",
                "responses": {
                    "200": {
                        "description": "Successful refresh.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JwtTokenResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/signout": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign out.",
                "description": "Sign the user out (Invalidate the token).",
                "operationId": "72aa8c73487b3660ba45c64c209aaa49",
                "responses": {
                    "200": {
                        "description": "Successful signout.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/current-user": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Current user.",
                "description": "It returns current user.",
                "operationId": "a72c6ebec0c6080156492ddaf6af82c5",
                "responses": {
                    "200": {
                        "description": "Successful signout.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/posts": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "List of posts.",
                "description": "It returns a list of posts.",
                "operationId": "792612e14feab734df8d4d9f8671d47b",
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostItem"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Add post.",
                "description": "It adds a new post to the DB.",
                "operationId": "d4e850d253397c85457b0c4bbd2fe0ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostItemAdd"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostItem"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/posts/{id}": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "One post.",
                "description": "It returns one post.",
                "operationId": "3680cae7eb31f337305a28d17d72a23d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostItem"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "posts"
                ],
                "summary": "Delete post.",
                "description": "It deletes the post from the DB.",
                "operationId": "e8bf457636aad55e305ac07826975008",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "posts"
                ],
                "summary": "Update post.",
                "description": "It updates the post in the DB.",
                "operationId": "2043f13dfc76cc825290a02927775303",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id.",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostItemUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostItem"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Success": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "type": "string",
                        "example": "string"
                    }
                },
                "type": "object"
            },
            "Error": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 400
                    },
                    "message": {
                        "type": "string",
                        "example": "string"
                    }
                },
                "type": "object"
            },
            "Unauthorized": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 401
                    },
                    "message": {
                        "type": "string",
                        "example": "string"
                    }
                },
                "type": "object"
            },
            "NotFoundError": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 404
                    },
                    "message": {
                        "type": "string",
                        "example": "string"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 422
                    },
                    "message": {
                        "type": "string",
                        "example": "string"
                    },
                    "errors": {
                        "type": "object",
                        "example": "{}"
                    }
                },
                "type": "object"
            },
            "ServerError": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 500
                    },
                    "message": {
                        "type": "string",
                        "example": "string"
                    }
                },
                "type": "object"
            },
            "SignupData": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John"
                    },
                    "email": {
                        "type": "string",
                        "example": "john-example@mail.mail"
                    },
                    "password": {
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
            "SigninData": {
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "john-example@mail.mail"
                    },
                    "password": {
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
            "JwtTokenResponse": {
                "properties": {
                    "token": {
                        "description": "Prepare JWT token response",
                        "type": "string",
                        "example": "string"
                    },
                    "expires_in": {
                        "type": "integer",
                        "example": 3600
                    }
                },
                "type": "object"
            },
            "PostItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Title"
                    },
                    "likes": {
                        "type": "integer",
                        "example": 15
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "04.10.2023 07:56:21"
                    }
                },
                "type": "object"
            },
            "PostItemAdd": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Title"
                    }
                },
                "type": "object"
            },
            "PostItemUpdate": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Title"
                    },
                    "likes": {
                        "type": "integer",
                        "example": 15
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}